# docs/api/api/openapi.yaml
openapi: 3.1.0
info:
  title: Courier Delivery Management System
  description: "\n        ## Courier Delivery Management System API\n\n        A comprehensive\
    \ API for managing courier delivery operations with real-time tracking and updates.\n\
    \n        ### Features\n        - **User Management**: Authentication and role-based\
    \ access control\n        - **Package Management**: Create, track, and manage\
    \ delivery packages\n        - **Real-time Updates**: WebSocket connections for\
    \ live notifications\n        - **Role-based Access**: Admin, Courier, Sender,\
    \ and Recipient roles\n        - **Status Tracking**: Complete package lifecycle\
    \ management\n\n        ### Authentication\n        This API uses JWT (JSON Web\
    \ Tokens) for authentication. Include the token in the Authorization header:\n\
    \        ```\n        Authorization: Bearer <your-jwt-token>\n        ```\n\n\
    \        ### User Roles\n        - **Admin**: Full system access and management\n\
    \        - **Courier**: Manage assigned packages and update delivery status\n\
    \        - **Sender**: Create packages and track deliveries\n        - **Recipient**:\
    \ View packages addressed to them\n\n        ### Getting Started\n        1. Register\
    \ a new user account\n        2. Login to receive an access token\n        3.\
    \ Use the token to access protected endpoints\n        4. Connect to WebSocket\
...
    \        "
  contact:
...
  license:
...
servers:
...
paths:
...
  /api/packages/:
    post:
      tags:
      - packages
      summary: Create Package
      description: Create a new package for delivery
      operationId: create_package_api_packages__post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '201':
          description: Package created successfully
          content:
            application/json:
              example:
                id: 1
                title: Important Documents
                description: Legal documents for delivery
                status: created
                sender_name: John Doe
                recipient_name: Jane Smith
                created_at: '2024-01-01T10:00:00Z'
        '401':
          description: Authentication required
          content:
            application/json:
              example:
                detail: Not authenticated
        '403':
          description: Insufficient permissions
          content:
            application/json:
              example:
                detail: 'Access denied. Required role: admin'
        '404':
          description: Resource not found
          content:
            application/json:
              example:
                detail: Resource not found
        '422':
          description: Validation error
          content:
            application/json:
              example:
                detail:
                - loc:
                  - body
                  - email
                  msg: field required
                  type: value_error.missing
        '500':
          description: Internal server error
          content:
            application/json:
              example:
                detail: Internal server error
    get:
      tags:
      - packages
      summary: List Packages
      description: Get list of packages based on user role and filters
      operationId: list_packages_api_packages__get
      security:
      - HTTPBearer: []
      parameters:
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          description: Number of packages to skip
          default: 0
          title: Skip
        description: Number of packages to skip
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Maximum number of packages to return
          default: 10
          title: Limit
        description: Maximum number of packages to return
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/PackageStatus'
          - type: 'null'
          description: Filter by package status
          title: Status
        description: Filter by package status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search packages by title, description, or names
          title: Search
        description: Search packages by title, description, or names
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageResponse'
                title: Response List Packages Api Packages  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/packages/available:
    get:
      tags:
      - packages
      summary: Get Available Packages
      description: Get packages available for courier assignment
      operationId: get_available_packages_api_packages_available_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PackageResponse'
                type: array
                title: Response Get Available Packages Api Packages Available Get
      security:
      - HTTPBearer: []
...
  /api/packages/stats:
    get:
      tags:
      - packages
      summary: Get Package Statistics
      description: Get package statistics based on user role
      operationId: get_package_stats_api_packages_stats_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageStatsResponse'
      security:
      - HTTPBearer: []
  /api/packages/{package_id}:
    get:
      tags:
      - packages
      summary: Get Package Details
      description: Get detailed information about a specific package
      operationId: get_package_api_packages__package_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          title: Package Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - packages
      summary: Update Package
      description: Update package information
      operationId: update_package_api_packages__package_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          title: Package Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - packages
      summary: Delete Package
      description: Delete a package (Admin or sender only)
      operationId: delete_package_api_packages__package_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          title: Package Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/packages/{package_id}/status:
    put:
      tags:
      - packages
      summary: Update Package Status
      description: Update package delivery status
      operationId: update_package_status_api_packages__package_id__status_put
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          title: Package Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/packages/{package_id}/assign:
    put:
      tags:
      - packages
      summary: Assign Package to Courier
      description: Assign package to a specific courier
      operationId: assign_package_api_packages__package_id__assign_put
      security:
      - HTTPBearer: []
      parameters:
      - name: package_id
        in: path
        required: true
        schema:
          type: integer
          title: Package Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageAssignment'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/packages/search/{search_term}:
    get:
      tags:
      - packages
      summary: Search Packages
      description: Search packages by various criteria
      operationId: search_packages_api_packages_search__search_term__get
      security:
      - HTTPBearer: []
      parameters:
      - name: search_term
        in: path
        required: true
        schema:
          type: string
          title: Search Term
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PackageResponse'
                title: Response Search Packages Api Packages Search  Search Term  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PackageAssignment:
      properties:
        courier_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Courier Id
          description: ID of the courier to assign
      type: object
      required:
      - courier_id
      title: PackageAssignment
      description: Schema for assigning package to courier.
    PackageCreate:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        sender_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Sender Name
        sender_phone:
          type: string
          maxLength: 20
          minLength: 1
          title: Sender Phone
        sender_address:
          type: string
          maxLength: 500
          minLength: 1
          title: Sender Address
        recipient_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Recipient Name
        recipient_phone:
          type: string
          maxLength: 20
          minLength: 1
          title: Recipient Phone
        recipient_address:
          type: string
          maxLength: 500
          minLength: 1
          title: Recipient Address
      type: object
      required:
      - title
      - sender_name
      - sender_phone
      - sender_address
      - recipient_name
      - recipient_phone
      - recipient_address
      title: PackageCreate
      description: Schema for creating a new package.
    PackageResponse:
      properties:
        title:
          type: string
          maxLength: 200
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        sender_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Sender Name
        sender_phone:
          type: string
          maxLength: 20
          minLength: 1
          title: Sender Phone
        sender_address:
          type: string
          maxLength: 500
          minLength: 1
          title: Sender Address
        recipient_name:
          type: string
          maxLength: 100
          minLength: 1
          title: Recipient Name
        recipient_phone:
          type: string
          maxLength: 20
          minLength: 1
          title: Recipient Phone
        recipient_address:
          type: string
          maxLength: 500
          minLength: 1
          title: Recipient Address
        id:
          type: integer
          title: Id
        status:
          $ref: '#/components/schemas/PackageStatus'
        sender_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sender Id
        courier_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Courier Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - title
      - sender_name
      - sender_phone
      - sender_address
      - recipient_name
      - recipient_phone
      - recipient_address
      - id
      - status
      - created_at
      - updated_at
      title: PackageResponse
      description: Schema for package response.
    PackageStatsResponse:
      properties:
        total_packages:
          type: integer
          title: Total Packages
        packages_by_status:
          additionalProperties:
            type: integer
          type: object
          title: Packages By Status
        recent_packages:
          type: integer
          title: Recent Packages
        user_role:
          type: string
          title: User Role
        total_couriers:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total Couriers
        unassigned_packages:
          anyOf:
          - type: integer
          - type: 'null'
          title: Unassigned Packages
      type: object
      required:
      - total_packages
      - packages_by_status
      - recent_packages
      - user_role
      title: PackageStatsResponse
      description: Schema for package statistics response.
    PackageStatus:
      type: string
      enum:
      - created
      - assigned
      - picked_up
      - in_transit
      - delivered
      - failed
      - cancelled
      title: PackageStatus
    PackageStatusUpdate:
      properties:
        status:
          allOf:
          - $ref: '#/components/schemas/PackageStatus'
          description: New package status
      type: object
      required:
      - status
      title: PackageStatusUpdate
      description: Schema for updating package status.
    PackageUpdate:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Description
        sender_name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Sender Name
        sender_phone:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Sender Phone
        sender_address:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Sender Address
        recipient_name:
          anyOf:
          - type: string
            maxLength: 100
            minLength: 1
          - type: 'null'
          title: Recipient Name
        recipient_phone:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 1
          - type: 'null'
          title: Recipient Phone
        recipient_address:
          anyOf:
          - type: string
            maxLength: 500
            minLength: 1
          - type: 'null'
          title: Recipient Address
      type: object
      title: PackageUpdate
      description: Schema for updating package information.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login endpoint
security:
- BearerAuth: []
...